+++ models/ExternalCalendarEntryQuery.php
@@ -4,7 +4,7 @@ namespace humhub\modules\external_calendar\models;
 use humhub\modules\calendar\interfaces\AbstractCalendarQuery;
 use humhub\modules\calendar\interfaces\CalendarItemWrapper;
 use humhub\modules\calendar\interfaces\VCalendar;
-use humhub\modules\external_calendar\CalendarUtils;
+use humhub\modules\external_calendar\helpers\CalendarUtils;
 use humhub\modules\user\models\User;
 use Sabre\VObject\Component\VEvent;
 use Sabre\VObject\Recur\EventIterator;